requirements.txt
google-api-python-client==1.7.8
cryptoauthlib
cryptography >= 2.3
paho-mqtt==1.4.0
